+++ plugins/postcss-custom-properties/CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ### Unreleased
 
+- Add `overrideImportFromWithRoot` option
+- Allow `.mjs` in `importFrom`
 - Converted to typescript
 - Correct typings for plugin options
-- Allow `.mjs` in `importFrom`
 - Fix unicode support in custom property names
 
 ### 12.0.4 (January 7, 2022)